Module H2

H2 is a high-performance, memory-efficient, and scalable HTTP/2 implementation for OCaml. It is based on the concepts in http/af, and therefore uses the Angstrom and Faraday libraries to implement the parsing and serialization layers of the HTTP/2 standard. It also preserves the same API as http/af wherever possible.

Not unlike http/af, the user should be familiar with HTTP, and the basic principles of memory management and vectorized IO in order to use this library.

Basic HTTP Types

module Method : module type of Httpaf.Method

Request Method

module Status : sig ... end

Response Status Codes

module Headers : sig ... end

Header Fields

Message Body

module Body : sig ... end

Message Types

module Request : sig ... end

Request

module Response : sig ... end

Response

module IOVec : module type of Httpaf.IOVec

IOVec

module Reqd : sig ... end

module Error_code : sig ... end

module Settings : sig ... end

module Config : sig ... end

Server Connection

module Server_connection : sig ... end

Client Connection

module Client_connection : sig ... end