Module H2.Response

Response

A server-generated message to a Request.t.

type t = {

status : Status.t;

headers : Headers.t;

}

val create : ?⁠headers:Headers.t -> Status.t -> t

create ?headers status creates an HTTP response with the given parameters. Unlike the Response type in http/af, h2 does not define a way for responses to carry reason phrases or protocol version.

See RFC75408.1.2.4 for more details.

val body_length : t -> [ `Error of [ `Bad_request ] | `Fixed of int64 | `Unknown ]

body_length t is the length of the message body accompanying t.

See RFC72303.3.3 for more details.

val pp_hum : Stdlib.Format.formatter -> t -> unit